requirements.txt
Flask==2.0.3
google-cloud-datastore==2.4.0
google-cloud-storage==2.1.0
google-auth==1.30.0
requests==2.27.1
